add LLM integration and change to the server
