+++ buf.gen.yaml
@@ -10,10 +10,10 @@ plugins:
       - paths=source_relative
       - require_unimplemented_servers=false
   - plugin: buf.build/bufbuild/es
-    out: client/ts/src/generated
+    out: client/src/generated
     opt:
       - target=ts
   - plugin: buf.build/connectrpc/es
-    out: client/ts/src/generated
+    out: client/src/generated
     opt:
       - target=ts
