+++ client/package.json
@@ -1,0 +1,48 @@
+{
+  "name": "mcp-grpc-client-ui",
+  "version": "0.1.0",
+  "description": "Unified MCP gRPC client with TypeScript SDK and React UI",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "private": true,
+  "scripts": {
+    "build": "vite build && tsc -p tsconfig.server.json && cp -r src/generated/buf dist/generated/",
+    "start": "node dist/server.js",
+    "dev": "vite",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "@bufbuild/protobuf": "^2.8.0",
+    "@bufbuild/protovalidate": "^1.0.0",
+    "@connectrpc/connect": "^2.1.0",
+    "@connectrpc/connect-node": "^2.1.0",
+    "@grpc/grpc-js": "^1.9.0",
+    "@grpc/proto-loader": "^0.7.10",
+    "@types/node": "^20.8.0",
+    "google-protobuf": "^3.21.2",
+    "jsonwebtoken": "^9.0.2",
+    "openai": "^5.23.0",
+    "protobufjs": "^7.5.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "@types/google-protobuf": "^3.15.12",
+    "@types/jsonwebtoken": "^9.0.3",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
+    "@vitejs/plugin-react": "^4.0.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.0"
+  },
+  "keywords": [
+    "mcp",
+    "grpc",
+    "protobuf",
+    "client",
+    "react",
+    "ui"
+  ],
+  "license": "MIT"
+}
