--- client/ts/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@strategific/mcp-grpc-client",
-  "version": "0.1.0",
-  "description": "TypeScript client SDK for gRPC MCP",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "dev": "ts-node src/examples/dynamic_call.ts",
-    "test-gateway": "ts-node src/examples/test_gateway.ts",
-    "test": "jest",
-    "ui": "node ../ui/working_server.js"
-  },
-  "dependencies": {
-    "@bufbuild/protobuf": "^2.8.0",
-    "@bufbuild/protovalidate": "^1.0.0",
-    "@connectrpc/connect": "^2.1.0",
-    "@connectrpc/connect-node": "^2.1.0",
-    "@grpc/grpc-js": "^1.9.0",
-    "@grpc/proto-loader": "^0.7.10",
-    "google-protobuf": "^3.21.2",
-    "jsonwebtoken": "^9.0.2",
-    "openai": "^5.23.0",
-    "protobufjs": "^7.5.4"
-  },
-  "devDependencies": {
-    "@types/google-protobuf": "^3.15.12",
-    "@types/jsonwebtoken": "^9.0.3",
-    "@types/node": "^20.8.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.2.2"
-  },
-  "keywords": [
-    "mcp",
-    "grpc",
-    "protobuf",
-    "client"
-  ],
-  "license": "MIT"
-}
