--- client/ui/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "mcp-grpc-ui",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "vite build && tsc -p tsconfig.server.json",
-    "start": "node dist/server.js",
-    "dev": "ts-node src/server.ts",
-    "vite": "vite",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@bufbuild/protobuf": "^2.8.0",
-    "@grpc/grpc-js": "^1.9.0",
-    "@grpc/proto-loader": "^0.7.10",
-    "@types/node": "^20.8.0",
-    "openai": "^5.23.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.0",
-    "@vitejs/plugin-react": "^4.0.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.0"
-  }
-}
