--- client/ui/tsconfig.server.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "ES2020",
-    "moduleResolution": "node",
-    "lib": ["ES2020", "DOM"],
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "declaration": false,
-    "sourceMap": true,
-    "allowJs": true
-  },
-  "include": ["src/server.ts"],
-  "exclude": ["node_modules", "dist", "src/main.tsx", "src/App.tsx"]
-}
